Implemented the usage of ConnectionResult in the TextFileConnection class. More below:
Removed all custom exceptions from use since their faulty result could be returned through ConnectionResult instead.

+++ PswManagerDatabase/DataAccess/TextFileConnection.cs
@@ -1,7 +1,6 @@
 ﻿using PswManagerDatabase.Config;
 using PswManagerDatabase.DataAccess.Interfaces;
 using PswManagerDatabase.DataAccess.TextFileConnHelper;
-using PswManagerDatabase.Exceptions;
 using PswManagerDatabase.Models;
 using PswManagerHelperMethods;
 using System;
@@ -12,6 +11,7 @@ using System.Text;
 using System.Threading.Tasks;
 
 namespace PswManagerDatabase.DataAccess {
+
     public class TextFileConnection : IDataConnection {
 
         readonly IPaths paths;
@@ -58,18 +58,22 @@ namespace PswManagerDatabase.DataAccess {
             return paths;
         }
 
-        public void CreateAccount(AccountModel model) {
-            if(SearchByName(model.Name) != null) {
-                throw new AccountExistsAlreadyException();
+        public ConnectionResult CreateAccount(AccountModel model) {
+            if(AccountExist(model.Name)) {
+                return new ConnectionResult(false, "The given account name is already occupied.");
             }
 
             File.AppendAllLines(paths.AccountsFilePath, new[] { model.Name });
             File.AppendAllLines(paths.PasswordsFilePath, new[] { model.Password });
             File.AppendAllLines(paths.EmailsFilePath, new[] { model.Email });
+
+            return new ConnectionResult(true);
         }
 
-        public AccountModel GetAccount(string name) {
-            int position = SearchByName(name) ?? throw new InexistentAccountException("The given account doesn't exist.");
+        public ConnectionResult<AccountModel> GetAccount(string name) {
+            if(!AccountExist(name, out int position)) {
+                return new ConnectionResult<AccountModel>(false, "The given account doesn't exist.");
+            }
 
             AccountModel output = new();
 
@@ -77,10 +81,10 @@ namespace PswManagerDatabase.DataAccess {
             output.Password = File.ReadAllLines(paths.PasswordsFilePath).Skip(position).Take(1).First();
             output.Email = File.ReadAllLines(paths.EmailsFilePath).Skip(position).Take(1).First();
 
-            return output;
+            return new ConnectionResult<AccountModel>(true, output);
         }
 
-        public List<AccountModel> GetAllAccounts() {
+        public ConnectionResult<List<AccountModel>> GetAllAccounts() {
 
             var names = File.ReadAllLines(paths.AccountsFilePath);
             var passwords = File.ReadAllLines(paths.PasswordsFilePath);
@@ -91,13 +95,21 @@ namespace PswManagerDatabase.DataAccess {
                 .Select(x => new AccountModel(names[x], passwords[x], emails[x]))
                 .ToList();
 
-            return accounts;
+            return new ConnectionResult<List<AccountModel>>(true, accounts);
         }
 
-        public AccountModel UpdateAccount(string name, AccountModel newModel) {
-            int position = SearchByName(name) ?? throw new InexistentAccountException("The given account doesn't exist.");
+        public ConnectionResult<AccountModel> UpdateAccount(string name, AccountModel newModel) {
 
-            (position < 0 || position > currentLength).IfTrueThrow(new InexistentAccountException("The given number is out of range."));
+            if(!AccountExist(name, out int position)) {
+                return new ConnectionResult<AccountModel>(false, "The given account doesn't exist.");
+            }
+
+            (position < 0 || position > currentLength).IfTrueThrow(
+                new ArgumentOutOfRangeException(
+                    nameof(name), 
+                    "The given name has been found in an out of range position. The files might be corrupted."
+                )
+            );
 
             EditValue(paths.AccountsFilePath, position, newModel.Name);
             EditValue(paths.PasswordsFilePath, position, newModel.Password);
@@ -106,7 +118,7 @@ namespace PswManagerDatabase.DataAccess {
             return GetAccount(newModel.Name ?? name);
         }
 
-        private void EditValue(string path, int position, string value) {
+        private static void EditValue(string path, int position, string value) {
             if(value is not null) {
                 var list = File.ReadAllLines(path);
                 list[position] = value;
@@ -115,15 +127,19 @@ namespace PswManagerDatabase.DataAccess {
             }
         }
 
-        public void DeleteAccount(string name) {
-            int position = SearchByName(name) ?? throw new InexistentAccountException("The given account doesn't exist.");
+        public ConnectionResult DeleteAccount(string name) {
+            if(!AccountExist(name, out int position)) {
+                return new ConnectionResult(false, "The given account doesn't exist.");
+            }
 
             DeleteValue(paths.AccountsFilePath, position);
             DeleteValue(paths.PasswordsFilePath, position);
             DeleteValue(paths.EmailsFilePath, position);
+
+            return new ConnectionResult(true);
         }
 
-        private void DeleteValue(string path, int position) {
+        private static void DeleteValue(string path, int position) {
             List<string> list = File.ReadAllLines(path).ToList();
             list.RemoveAt(position);
             File.WriteAllLines(path, list);
@@ -131,6 +147,21 @@ namespace PswManagerDatabase.DataAccess {
 
         public bool AccountExist(string name) {
             return File.Exists(paths.AccountsFilePath) && SearchByName(name) != null;
+        }
+
+        private bool AccountExist(string name, out int position) {
+            position = -1;
+
+            if(!File.Exists(paths.AccountsFilePath))
+                return false;
+
+            int? temp = SearchByName(name);
+            if(temp == null)
+                return false;
+
+            position = (int)temp;
+
+            return true;
         }
     }
 }
